@@ -1,36 +1,43 @@
 {
-  "name": "unique-wallet",
-  "repository": "https://github.com/usetech-llc/wallet",
+  "name": "unique-network-marketplace",
+  "repository": "https://github.com/UniqueNetwork/unique-wallet",
   "main": "packages/apps-electron/build/electron.js",
-  "description": "A wallet for unique nft tokens",
-  "author": "Jaco Greeff <jacogr@gmail.com>, Konstantin Pozdnikin <kpozdnikin@usetech.ru>",
-  "version": "1.0.0",
+  "description": "A wallet for unique NFT tokens",
+  "author": "Jaco Greeff <jacogr@gmail.com>, Konstantin Pozdnikin <kpozdnikin@usetech.com>",
+  "version": "0.83.2-28",
   "license": "Apache-2",
   "homepage": ".",
+  "private": true,
   "workspaces": [
     "packages/*"
   ],
   "resolutions": {
-    "@polkadot/api": "^4.15.1",
-    "@polkadot/api-contract": "^4.15.1",
-    "@polkadot/hw-ledger": "^6.9.1",
-    "@polkadot/keyring": "^6.9.1",
-    "@polkadot/networks": "^6.9.1",
-    "@polkadot/phishing": "^0.6.203",
-    "@polkadot/types": "^4.15.1",
-    "@polkadot/util": "^6.9.1",
-    "@polkadot/util-crypto": "^6.9.1",
-    "@polkadot/wasm-crypto": "^4.0.2",
-    "@polkadot/x-rxjs": "^6.9.1",
+    "@polkadot/api": "^5.5.1",
+    "@polkadot/api-contract": "^5.5.1",
+    "@polkadot/hw-ledger": "^7.2.1",
+    "@polkadot/keyring": "^7.2.1",
+    "@polkadot/networks": "^7.2.1",
+    "@polkadot/phishing": "^0.6.257",
+    "@polkadot/types": "^5.5.1",
+    "@polkadot/types-known": "^5.5.1",
+    "@polkadot/types-support": "^5.5.1",
+    "@polkadot/util": "^7.2.1",
+    "@polkadot/util-crypto": "^7.2.1",
+    "@polkadot/wasm-crypto": "^4.1.2",
     "babel-core": "^7.0.0-bridge.0",
     "styled-components": "^5.2.0",
-    "typescript": "^4.3.2"
+    "tar": "^6.1.6",
+    "typescript": "^4.3.5",
+    "webpack": "^5.50.0",
+    "webpack-cli": "^4.8.0"
   },
   "scripts": {
     "analyze": "yarn clean && BUILD_ANALYZE=1 yarn run build:code && yarn source-map-explorer packages/apps/build/main.*.js",
+    "audit": "yarn npm audit --recursive --all",
+    "audit:ci": "yarn npm audit --recursive --all --severity high",
     "build": "yarn run build:i18n && yarn run build:code",
     "build:code": "NODE_ENV=production yarn polkadot-dev-build-ts",
-    "build:i18n": "i18next-scanner --config i18next-scanner.config.cjs && node ./scripts/i18nSort.cjs",
+    "build:i18n": "i18next-scanner --config i18next-scanner.config.js && node ./scripts/i18nSort.cjs",
     "build:release:electron": "yarn build && yarn build:electron && yarn postinstall:electron",
     "build:release:ipfs": "node scripts/ipfsUpload.cjs",
     "build:release:ghpages": "yarn polkadot-ci-ghact-docs",
