@@ -11,20 +11,20 @@
   "contributors": [],
   "license": "Apache-2.0",
   "dependencies": {
-    "@babel/core": "^7.14.3",
+    "@babel/core": "^7.15.0",
     "@babel/polyfill": "^7.12.1",
-    "@babel/runtime": "^7.14.0",
+    "@babel/runtime": "^7.15.3",
     "@polkadot/apps-config": "^0.83.2-28",
     "@polkadot/apps-routing": "^0.83.2-28",
-    "@polkadot/dev": "^0.62.5",
-    "@polkadot/react-api": "^0.91.3-0",
+    "@polkadot/dev": "^0.62.60",
+    "@polkadot/react-api": "^0.94.2-95",
     "@polkadot/react-components": "^0.83.2-28",
     "@polkadot/react-hooks": "^0.83.2-28",
     "@polkadot/react-params": "^0.83.2-28",
     "@polkadot/react-query": "^0.83.2-28",
     "@polkadot/react-signer": "^0.83.2-28",
     "punycode": "^2.1.1",
-    "query-string": "^7.0.0"
+    "query-string": "^7.0.1"
   },
   "devDependencies": {
     "@types/punycode": "^2.1.0",
@@ -32,21 +32,21 @@
     "buffer": "^6.0.3",
     "copy-webpack-plugin": "^7.0.0",
     "crypto-browserify": "^3.12.0",
-    "css-loader": "^5.2.6",
+    "css-loader": "^5.2.7",
     "empty": "^0.10.1",
     "file-loader": "^6.2.0",
     "html-loader": "^1.3.2",
     "html-webpack-plugin": "^4.5.2",
     "markdown-loader": "^6.0.0",
-    "mini-css-extract-plugin": "^1.6.0",
+    "mini-css-extract-plugin": "^1.6.2",
     "null-loader": "^4.0.1",
     "path-browserify": "^1.0.1",
     "process": "^0.11.10",
     "stream-browserify": "^3.0.0",
     "url-loader": "^4.1.1",
-    "webpack": "^5.38.1",
-    "webpack-cli": "^4.7.0",
-    "webpack-dev-server": "^4.0.0-beta.0",
-    "webpack-merge": "^5.7.3"
+    "webpack": "^5.50.0",
+    "webpack-cli": "^4.8.0",
+    "webpack-dev-server": "^4.0.0-rc.0",
+    "webpack-merge": "^5.8.0"
   }
 }
