@@ -3,13 +3,12 @@
 
 // TODO: Lots of duplicated code between this and withObservable, surely there is a better way of doing this?
 
+import type { Observable } from 'rxjs';
 import type { CallState } from '../types';
 import type { DefaultProps, HOC, Options, RenderFn } from './types';
 
 import React from 'react';
-
-import { Observable, of } from '@polkadot/x-rxjs';
-import { catchError, map } from '@polkadot/x-rxjs/operators';
+import { catchError, map, of } from 'rxjs';
 
 import echoTransform from '../transform/echo';
 import { intervalObservable, isEqual, triggerChange } from '../util';
@@ -23,14 +22,14 @@ export default function withObservable<T, P> (observable: Observable<P>, { callO
     class WithObservable extends React.Component<any, State> {
       private isActive = true;
 
-      public state: State = {
+      public override state: State = {
         callResult: undefined,
         callUpdated: false,
         callUpdatedAt: 0,
         subscriptions: []
       };
 
-      public componentDidMount (): void {
+      public override componentDidMount (): void {
         this.setState({
           subscriptions: [
             observable
@@ -44,7 +43,7 @@ export default function withObservable<T, P> (observable: Observable<P>, { callO
         });
       }
 
-      public componentWillUnmount (): void {
+      public override componentWillUnmount (): void {
         this.isActive = false;
         this.state.subscriptions.forEach((subscription): void =>
           subscription.unsubscribe()
@@ -69,7 +68,7 @@ export default function withObservable<T, P> (observable: Observable<P>, { callO
         }
       }
 
-      public render (): React.ReactNode {
+      public override render (): React.ReactNode {
         const { children } = this.props;
         // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
         const { callResult, callUpdated, callUpdatedAt } = this.state;
