@@ -3,11 +3,11 @@
 
 import BN from 'bn.js';
 
-import networks from '@polkadot/networks';
+import { selectableNetworks } from '@polkadot/networks';
 import { assert } from '@polkadot/util';
 
 function getGenesis (name: string): string {
-  const network = networks.find(({ network }) => network === name);
+  const network = selectableNetworks.find(({ network }: { network: string }) => network === name);
 
   assert(network && network.genesisHash[0], `Unable to find genesisHash for ${name}`);
 
@@ -21,4 +21,8 @@ export const KUSAMA_GENESIS = getGenesis('kusama');
 export const POLKADOT_GENESIS = getGenesis('polkadot');
 export const POLKADOT_DENOM_BLOCK = new BN(1248328);
 
-export const ROCOCO_GENESIS = '0x47381ee0697153d64404fc578392c8fd5cba9073391908f46c888498415647bd';
+export const ROCOCO_GENESIS = '0xe7c3d5edde7db964317cd9b51a3a059d7cd99f81bdbce14990047354334c9779';
+
+export const WESTEND_GENESIS = '0xe143f23803ac50e8f6f8e62695d1ce9e4e1d68aa36c1cd2cfd15340213f3423e';
+
+export const NEATCOIN_GENESIS = '0xfbb541421d30423c9a753ffa844b64fd44d823f513bf49e3b73b3a656309a595';
