@@ -8,7 +8,7 @@ import envConfig from '@polkadot/apps-config/envConfig';
 
 import { expandEndpoints } from './util';
 
-const { wssUrl } = envConfig;
+const { uniqueSubstrateApi } = envConfig;
 /* eslint-disable sort-keys */
 
 // The available endpoints that will show in the dropdown. For the most part (with the exception of
@@ -240,7 +240,7 @@ export function createTesting (t: TFunction): LinkOption[] {
       info: 'unique',
       text: t('rpc.unique', 'Unique', { ns: 'apps-config' }),
       providers: {
-        Unique: wssUrl || 'wss://testnet2.uniquenetwork.io'
+        Unique: uniqueSubstrateApi
       }
     },
     {
