@@ -4,13 +4,19 @@
 import type { TFunction } from 'i18next';
 import type { Option } from './types';
 
-import known from '@polkadot/networks';
+import { selectableNetworks } from '@polkadot/networks';
 
-const networks = known.map(({ displayName, network, prefix }) => ({
-  info: network,
-  text: displayName,
-  value: prefix
-}));
+const networks = selectableNetworks
+  .map(({ displayName, network, prefix }) => ({
+    info: network,
+    text: displayName,
+    value: prefix
+  }))
+  .sort((a, b) =>
+    [0, 2, 42].includes(a.value) || [0, 2, 42].includes(b.value)
+      ? 0
+      : a.text.localeCompare(b.text)
+  );
 
 // Definitions here are with the following values -
 //   info: the name of a logo as defined in ../logos, specifically in namedLogos
