@@ -2,10 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { TFunction } from 'i18next';
-
-import Component from '@polkadot/app-nft-wallet';
+import React from 'react';
 
 import { Route } from './types';
+
+// import Component from '@polkadot/app-nft-wallet';
+
+const Component = React.lazy(() => import('@polkadot/app-nft-wallet'));
 
 export default function create (t: TFunction): Route {
   return {
