@@ -8,6 +8,8 @@ const { merge } = require('webpack-merge');
 
 const baseConfig = require('./webpack.base.cjs');
 
+require('dotenv-defaults').config();
+
 module.exports = merge(
   baseConfig(__dirname, 'development'),
   {
@@ -17,27 +19,27 @@ module.exports = merge(
       proxy: {
         '/health': {
           changeOrigin: true,
-          target: 'https://unqnft.io/'
+          target: process.env.UNIQUE_API
         },
         '/mint': {
           changeOrigin: true,
-          target: 'https://unqnft.io/'
+          target: process.env.UNIQUE_API
         },
         '/offers': {
           changeOrigin: true,
-          target: 'https://unqnft.io/'
+          target: process.env.UNIQUE_API
         },
         '/trades': {
           changeOrigin: true,
-          target: 'https://unqnft.io/'
+          target: process.env.UNIQUE_API
         },
-        'https://unqnft.io': {
+        [process.env.UNIQUE_API]: {
           changeOrigin: true,
-          target: 'https://unqnft.io/'
+          target: process.env.UNIQUE_API
         },
-        'https://whitelabel.market': {
+        [process.env.WHITE_LABEL_URL]: {
           changeOrigin: true,
-          target: 'https://whitelabel.market/'
+          target: process.env.WHITE_LABEL_URL
         }
       },
       static: path.resolve(__dirname, 'build')
