@@ -8,23 +8,19 @@ import type { HoldType } from '@polkadot/react-hooks/useCollections';
 
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import InfiniteScroll from 'react-infinite-scroller';
-import { useHistory } from 'react-router';
 import Confirm from 'semantic-ui-react/dist/commonjs/addons/Confirm';
-import Button from 'semantic-ui-react/dist/commonjs/elements/Button/Button';
 import Loader from 'semantic-ui-react/dist/commonjs/elements/Loader';
 import Item from 'semantic-ui-react/dist/commonjs/views/Item';
 
 import envConfig from '@polkadot/apps-config/envConfig';
 import { Expander } from '@polkadot/react-components';
-import pencil from '@polkadot/react-components/ManageCollection/pencil.svg';
 import trash from '@polkadot/react-components/ManageCollection/trash.svg';
 import Tooltip from '@polkadot/react-components/Tooltip';
-import { useCollections, useDecoder, useMetadata } from '@polkadot/react-hooks';
-import { useCollection } from '@polkadot/react-hooks/useCollection';
+import { useDecoder, useMetadata, useMyTokens } from '@polkadot/react-hooks';
 
 import NftTokenCard from '../NftTokenCard';
 
-const { canCreateToken, canEditCollection, uniqueCollectionIds } = envConfig;
+const { uniqueCollectionIds } = envConfig;
 
 interface Props {
   account?: string;
@@ -33,26 +29,22 @@ interface Props {
   removeCollection: (collection: string) => void;
   onHold: HoldType[];
   openTransferModal: (collection: NftCollectionInterface, tokenId: string, balance: number) => void;
-  shouldUpdateTokens: string | undefined;
   tokensSelling: string[];
 }
 
 const perPage = 5;
 
-function NftCollectionCard ({ account, canTransferTokens, collection, onHold, openTransferModal, removeCollection, shouldUpdateTokens, tokensSelling }: Props): React.ReactElement<Props> {
+function NftCollectionCard ({ account, canTransferTokens, collection, onHold, openTransferModal, removeCollection, tokensSelling }: Props): React.ReactElement<Props> {
   const [opened, setOpened] = useState(true);
+  const [currentPerPage, setCurrentPerPage] = useState(5);
   const [collectionImageUrl, setCollectionImageUrl] = useState<string>();
-  const [ownTokensCount, setOwnTokensCount] = useState<number>();
-  const [allTokensCount, setAllTokensCount] = useState<number>();
   const [confirmDeleteCollection, setConfirmDeleteCollection] = useState<boolean>(false);
-  const [tokensOnPage, setTokensOnPage] = useState<string[]>([]);
-  const [allMyTokens, setAllMyTokens] = useState<string[]>([]);
-  const { getTokensOfCollection } = useCollections();
-  const { getCollectionTokensCount } = useCollection();
   const { collectionName16Decoder } = useDecoder();
   const cleanup = useRef<boolean>(false);
-  const history = useHistory();
   const { getTokenImageUrl } = useMetadata();
+  const { allMyTokens, allTokensCount, ownTokensCount, tokensOnPage } = useMyTokens(account, collection, onHold, tokensSelling, currentPerPage);
+  const nftWalletPanel = useRef<HTMLDivElement>(null);
+
   const hasMore = tokensOnPage.length < allMyTokens.length;
 
   const openCollection = useCallback((isOpen) => {
@@ -69,46 +61,6 @@ function NftCollectionCard ({ account, canTransferTokens, collection, onHold, op
     setCollectionImageUrl(collectionImage);
   }, [collection, getTokenImageUrl]);
 
-  const editCollection = useCallback((collectionId: string) => {
-    history.push(`/wallet/manage-collection?collectionId=${collectionId}`);
-  }, [history]);
-
-  const createToken = useCallback((collectionId: string) => {
-    history.push(`/wallet/manage-token?collectionId=${collectionId}`);
-  }, [history]);
-
-  const getTokensCount = useCallback(async () => {
-    if (!collection) {
-      return;
-    }
-
-    const tokensCount: number = await getCollectionTokensCount(collection.id) as number;
-
-    if (cleanup.current) {
-      return;
-    }
-
-    setAllTokensCount(parseFloat(tokensCount.toString()));
-  }, [collection, getCollectionTokensCount]);
-
-  const updateTokens = useCallback(async () => {
-    if (!account) {
-      return;
-    }
-
-    const tokens = (await getTokensOfCollection(collection.id, account)) as string[];
-
-    if (cleanup.current) {
-      return;
-    }
-
-    setOwnTokensCount(tokens.length);
-    const holdingTokens = onHold.map((item) => item.tokenId);
-
-    setAllMyTokens([...tokensSelling, ...tokens, ...holdingTokens]);
-    setTokensOnPage([...tokensSelling, ...tokens, ...holdingTokens].slice(0, perPage));
-  }, [account, collection, getTokensOfCollection, onHold, tokensSelling]);
-
   const toggleConfirmation = useCallback((status, e: React.MouseEvent<any>) => {
     e.stopPropagation();
 
@@ -116,18 +68,18 @@ function NftCollectionCard ({ account, canTransferTokens, collection, onHold, op
   }, []);
 
   const loadMore = useCallback((page: number) => {
-    setTokensOnPage(allMyTokens.slice(0, perPage * page));
-  }, [allMyTokens]);
+    // handle load more on scroll action
+    setCurrentPerPage(page * perPage);
+  }, []);
 
-  useEffect(() => {
-    if (shouldUpdateTokens && shouldUpdateTokens === collection.id) {
-      void updateTokens();
+  // set scroll parent to initialize scroll container in mobile or desktop
+  const getScrollParent = useCallback(() => {
+    if (nftWalletPanel.current && nftWalletPanel.current.offsetWidth > 1024) {
+      return nftWalletPanel.current;
     }
-  }, [collection.id, shouldUpdateTokens, updateTokens]);
 
-  useEffect(() => {
-    void updateTokens();
-  }, [account, updateTokens]);
+    return null;
+  }, [nftWalletPanel]);
 
   useEffect(() => {
     if (!collectionImageUrl && collection) {
@@ -136,16 +88,12 @@ function NftCollectionCard ({ account, canTransferTokens, collection, onHold, op
   }, [collection, collectionImageUrl, defineCollectionImage]);
 
   useEffect(() => {
-    if (collection && allTokensCount === undefined) {
-      void getTokensCount();
-    }
-  }, [allTokensCount, collection, getTokensCount]);
-
-  useEffect(() => {
     return () => {
       cleanup.current = true;
     };
   }, []);
+
+  const useWindow = !getScrollParent();
 
   return (
     <Expander
@@ -173,39 +121,11 @@ function NftCollectionCard ({ account, canTransferTokens, collection, onHold, op
                 <div className='collection-description'>{collectionName16Decoder(collection.Description)}</div>
               )}
             </div>
-            { canCreateToken && (
-              <Button
-                className='create-button'
-                onClick={createToken.bind(null, collection.id)}
-                primary
-              >
-                Create token
-              </Button>
-            )}
           </div>
           <div className='tokens-count'>
             <span>Total: {allTokensCount} {!allTokensCount || allTokensCount > 1 ? 'items' : 'item'} (own: {ownTokensCount || 0}, selling: {tokensSelling.length}, on hold: {onHold.length})</span>
           </div>
           <div className='link-button'>
-            { canEditCollection && (
-              <div className='link-button-with-tooltip'>
-                <img
-                  alt='edit'
-                  data-for='Edit collection'
-                  data-tip='Edit collection'
-                  onClick={editCollection.bind(null, collection.id)}
-                  src={pencil as string}
-                />
-                <Tooltip
-                  arrowColor={'transparent'}
-                  backgroundColor={'var(--border-color)'}
-                  place='bottom'
-                  text={'Edit collection'}
-                  textColor={'var(--sub-header-text-transform)'}
-                  trigger={'Edit collection'}
-                />
-              </div>
-            )}
             { !uniqueCollectionIds.includes(collection.id) && (
               <div className='link-button-with-tooltip'>
                 <img
@@ -236,8 +156,12 @@ function NftCollectionCard ({ account, canTransferTokens, collection, onHold, op
         </div>
       }
     >
-      <div className='expander-inner'>
+      <div
+        className='expander-inner'
+        ref={nftWalletPanel}
+      >
         <InfiniteScroll
+          getScrollParent={getScrollParent}
           hasMore={hasMore}
           initialLoad={false}
           loadMore={loadMore}
@@ -249,18 +173,17 @@ function NftCollectionCard ({ account, canTransferTokens, collection, onHold, op
           />}
           pageStart={1}
           threshold={100}
-          useWindow={false}
+          useWindow={useWindow}
         >
           <div className='token-table'>
-            { account && tokensOnPage.map((token) => (
+            { account && tokensOnPage.map((token: string, index: number) => (
               <NftTokenCard
                 account={account}
                 canTransferTokens={canTransferTokens}
                 collection={collection}
-                key={token}
+                key={`${token}-${index}`}
                 onHold={onHold}
                 openTransferModal={openTransferModal}
-                shouldUpdateTokens={shouldUpdateTokens}
                 token={token}
                 tokensSelling={tokensSelling}
               />
