@@ -5,31 +5,25 @@ import './styles.scss';
 
 // external imports
 import React, { useCallback, useEffect, useState } from 'react';
-import { Route, Switch, useHistory } from 'react-router';
+import { Route, Switch } from 'react-router';
 import { useLocation } from 'react-router-dom';
-import Button from 'semantic-ui-react/dist/commonjs/elements/Button/Button';
 import Header from 'semantic-ui-react/dist/commonjs/elements/Header/Header';
 
 import envConfig from '@polkadot/apps-config/envConfig';
-import { ManageCollection, ManageTokenAttributes, NftDetails } from '@polkadot/react-components';
+import { NftDetails } from '@polkadot/react-components';
 // local imports and components
 import { AppProps as Props } from '@polkadot/react-components/types';
 import { NftCollectionInterface } from '@polkadot/react-hooks/useCollection';
 
 import NftWallet from './containers/NftWallet';
 
-const { canAddCollections, canCreateCollection } = envConfig;
+const { canAddCollections } = envConfig;
 
-function App ({ account, basePath }: Props): React.ReactElement<Props> {
+function PageNftWallet ({ account, basePath, openPanel, setOpenPanel }: Props): React.ReactElement<Props> {
   const location = useLocation();
-  const history = useHistory();
   const [shouldUpdateTokens, setShouldUpdateTokens] = useState<string>();
   const collectionsStorage: NftCollectionInterface[] = JSON.parse(localStorage.getItem('tokenCollections') || '[]') as NftCollectionInterface[];
   const [collections, setCollections] = useState<NftCollectionInterface[]>(collectionsStorage);
-
-  const createCollection = useCallback(() => {
-    history.push('/wallet/manage-collection');
-  }, [history]);
 
   const addCollection = useCallback((collection: NftCollectionInterface) => {
     setCollections((prevCollections: NftCollectionInterface[]) => {
@@ -49,7 +43,6 @@ function App ({ account, basePath }: Props): React.ReactElement<Props> {
     const newCollectionList = collections.filter((item: NftCollectionInterface) => item.id !== collectionToRemove);
 
     setCollections(newCollectionList);
-
     localStorage.setItem('tokenCollections', JSON.stringify(newCollectionList));
   }, [collections]);
 
@@ -66,52 +59,23 @@ function App ({ account, basePath }: Props): React.ReactElement<Props> {
         <>
           <Header as='h1'>My Tokens</Header>
           <Header as='h4'>NFTs owned by me</Header>
-          { canCreateCollection && (
-            <Button
-              className='create-button'
-              onClick={createCollection}
-              primary
-            >
-              Create collection
-            </Button>
-          )}
         </>
       )}
       <Switch>
         <Route path={`${basePath}/token-details`}>
           <NftDetails
             account={account || ''}
-            setShouldUpdateTokens={setShouldUpdateTokens}
           />
         </Route>
-        <Route path={`${basePath}/manage-collection`}>
-          <ManageCollection
-            account={account}
-            addCollection={addCollection}
-            basePath={`${basePath}/manage-collection`}
-            setShouldUpdateTokens={setShouldUpdateTokens}
-          />
-        </Route>
-        <Route path={`${basePath}/manage-token`}>
-          <ManageTokenAttributes
-            account={account}
-            setShouldUpdateTokens={setShouldUpdateTokens}
-          />
-        </Route>
-        {/* <Route path={`${basePath}/tokens-for-sale`}>
-          <TokensForSale
-            account={account}
-            setShouldUpdateTokens={setShouldUpdateTokens}
-            shouldUpdateTokens={shouldUpdateTokens}
-          />
-        </Route> */}
         <Route path={basePath}>
           <NftWallet
             account={account}
             addCollection={addCollection}
             collections={collections}
+            openPanel={openPanel}
             removeCollectionFromList={removeCollectionFromList}
             setCollections={setCollections}
+            setOpenPanel={setOpenPanel}
             setShouldUpdateTokens={setShouldUpdateTokens}
             shouldUpdateTokens={shouldUpdateTokens}
           />
@@ -121,4 +85,4 @@ function App ({ account, basePath }: Props): React.ReactElement<Props> {
   );
 }
 
-export default React.memo(App);
+export default React.memo(PageNftWallet);
