@@ -1,12 +1,12 @@
 // Copyright 2017-2021 @polkadot/react-components authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ColumnProps } from './types';
+import type { ColumnsProps } from './types';
 
 import React from 'react';
 import styled from 'styled-components';
 
-function Columns ({ children, className = '' }: ColumnProps): React.ReactElement<ColumnProps> {
+function Columns ({ children, className = '' }: ColumnsProps): React.ReactElement<ColumnsProps> {
   return (
     <div className={`ui--Modal-Columns ${className}`}>{children}</div>
   );
