@@ -7,7 +7,6 @@ import BN from 'bn.js';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 // import Form from 'semantic-ui-react/dist/commonjs/collections/Form';
-import Grid from 'semantic-ui-react/dist/commonjs/collections/Grid';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Header from 'semantic-ui-react/dist/commonjs/elements/Header';
 import Image from 'semantic-ui-react/dist/commonjs/elements/Image';
@@ -15,6 +14,7 @@ import Loader from 'semantic-ui-react/dist/commonjs/elements/Loader';
 
 import envConfig from '@polkadot/apps-config/envConfig';
 import { TransferModal } from '@polkadot/react-components';
+import formatPrice from '@polkadot/react-components/util/formatPrice';
 import { useBalance, useDecoder, useMarketplaceStages, useSchema } from '@polkadot/react-hooks';
 
 import BuySteps from './BuySteps';
@@ -25,10 +25,9 @@ const { kusamaDecimals, showMarketActions } = envConfig;
 
 interface NftDetailsProps {
   account: string;
-  setShouldUpdateTokens?: (collectionId: string) => void;
 }
 
-function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React.ReactElement<NftDetailsProps> {
+function NftDetails ({ account }: NftDetailsProps): React.ReactElement<NftDetailsProps> {
   const query = new URLSearchParams(useLocation().search);
   const tokenId = query.get('tokenId') || '';
   const collectionId = query.get('collectionId') || '';
@@ -39,7 +38,7 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
   const { hex2a } = useDecoder();
   const { attributes, collectionInfo, reFungibleBalance, tokenUrl } = useSchema(account, collectionId, tokenId);
   const [tokenPriceForSale, setTokenPriceForSale] = useState<string>('');
-  const { cancelStep, deposited, escrowAddress, formatKsmBalance, getFee, getKusamaTransferFee, kusamaBalance, readyToAskPrice, sendCurrentUserAction, setPrice, setReadyToAskPrice, tokenAsk, tokenDepositor, tokenInfo, transferStep } = useMarketplaceStages(account, collectionInfo, tokenId);
+  const { cancelStep, deposited, escrowAddress, formatKsmBalance, getFee, getKusamaTransferFee, kusamaAvailableBalance, readyToAskPrice, sendCurrentUserAction, setPrice, setReadyToAskPrice, tokenAsk, tokenDepositor, tokenInfo, transferStep } = useMarketplaceStages(account, collectionInfo, tokenId);
 
   const uOwnIt = tokenInfo?.Owner?.toString() === account || (tokenAsk && tokenAsk.owner === account);
   const uSellIt = tokenAsk && tokenAsk.owner === account;
@@ -50,9 +49,8 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
 
   const goBack = useCallback((e: React.MouseEvent<HTMLAnchorElement>) => {
     e.preventDefault();
-    setShouldUpdateTokens && setShouldUpdateTokens('all');
     history.back();
-  }, [setShouldUpdateTokens]);
+  }, []);
 
   const onSavePrice = useCallback(() => {
     const parts = tokenPriceForSale.split('.');
@@ -66,8 +64,7 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
   const onTransferSuccess = useCallback(() => {
     setShowTransferForm(false);
     sendCurrentUserAction('UPDATE_TOKEN_STATE');
-    setShouldUpdateTokens && setShouldUpdateTokens(collectionId);
-  }, [collectionId, sendCurrentUserAction, setShouldUpdateTokens]);
+  }, [sendCurrentUserAction]);
 
   const closeAskModal = useCallback(() => {
     setReadyToAskPrice(false);
@@ -85,12 +82,16 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
       if (kusamaFees) {
         setKusamaFees(kusamaFees);
         const balanceNeeded = tokenAsk.price.add(getFee(tokenAsk.price)).add(kusamaFees.muln(2));
-        const low = !!kusamaBalance?.free.add(deposited || new BN(0)).lte(balanceNeeded);
+        const isLow = !!kusamaAvailableBalance?.add(deposited || new BN(0)).lte(balanceNeeded);
 
-        setLowKsmBalanceToBuy(low);
+        setLowKsmBalanceToBuy(isLow);
       }
     }
-  }, [deposited, escrowAddress, getFee, getKusamaTransferFee, kusamaBalance, tokenAsk]);
+  }, [deposited, escrowAddress, kusamaAvailableBalance, getFee, getKusamaTransferFee, tokenAsk]);
+
+  const getMarketPrice = useCallback((price: BN) => {
+    return formatPrice(formatKsmBalance(new BN(price).add(getFee(price))));
+  }, [formatKsmBalance, getFee]);
 
   useEffect(() => {
     void ksmFeesCheck();
@@ -119,24 +120,24 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
         </svg>
         back
       </a>
-      <Grid className='token-info'>
-        { (!collectionInfo || (account && (!kusamaBalance || !balance))) && (
+      <div className='token-info'>
+        { (!collectionInfo || (account && (!kusamaAvailableBalance || !balance))) && (
           <Loader
             active
             className='load-info'
             inline='centered'
           />
         )}
-        <Grid.Row>
-          <Grid.Column width={8}>
+        <div className='token-info--row'>
+          <div className='token-info--row--image'>
             { collectionInfo && (
               <Image
                 className='token-image-big'
                 src={tokenUrl}
               />
             )}
-          </Grid.Column>
-          <Grid.Column width={8}>
+          </div>
+          <div className='token-info--row--attributes'>
             <Header as='h3'>
               {collectionInfo && <span>{hex2a(collectionInfo.TokenPrefix)}</span>} #{tokenId}
             </Header>
@@ -157,7 +158,7 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
             { (tokenAsk && tokenAsk.price) && (
               <>
                 <Header as={'h2'}>
-                  {formatKsmBalance(tokenAsk.price.add(getFee(tokenAsk.price)))} KSM
+                  {getMarketPrice(tokenAsk.price)} KSM
                 </Header>
                 <p>Fee: {formatKsmBalance(getFee(tokenAsk.price))} KSM, Price: {formatKsmBalance(tokenAsk.price)} KSM</p>
                 {/* { (!uOwnIt && !transferStep && tokenAsk) && lowBalanceToBuy && (
@@ -207,8 +208,6 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
 
                 </div>
               )}
-              {console.log(showMarketActions, 'hhhhhhhhhhhhhh')}
-
               {showMarketActions && (
                 <>
                   { (!uOwnIt && !transferStep && tokenAsk && kusamaFees) && (
@@ -218,7 +217,6 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
                       <Button
                         content={`Buy it - ${formatKsmBalance(tokenAsk.price.add(getFee(tokenAsk.price)).add(kusamaFees.muln(2)))} KSM`}
                         disabled={ lowKsmBalanceToBuy}
-
                         onClick={sendCurrentUserAction.bind(null, 'BUY')}
                       />
                     </>
@@ -267,9 +265,9 @@ function NftDetails ({ account, setShouldUpdateTokens }: NftDetailsProps): React
               <BuySteps step={transferStep - 3} />
             )}
 
-          </Grid.Column>
-        </Grid.Row>
-      </Grid>
+          </div>
+        </div>
+      </div>
       { readyToAskPrice && (
         <SetPriceModal
           closeModal={closeAskModal}
