@@ -4,30 +4,44 @@
 import type { DeriveBalancesAll } from '@polkadot/api-derive/types';
 
 import BN from 'bn.js';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
 import { useApi, useCall, useKusamaApi } from '@polkadot/react-hooks';
 
 interface UseBalancesInterface {
-  balancesAll: DeriveBalancesAll | undefined;
-  deposited: BN | undefined;
-  kusamaBalancesAll: DeriveBalancesAll | undefined;
+  freeBalance: BN | undefined;
+  freeKusamaBalance: BN | undefined;
 }
 
-export const useBalances = (account: string | undefined, deposited: BN | undefined, getUserDeposit: () => Promise<BN | null>): UseBalancesInterface => {
+export const useBalances = (account: string | undefined, getUserDeposit?: () => Promise<BN | null>): UseBalancesInterface => {
   const { api } = useApi();
   const { kusamaApi } = useKusamaApi(account || '');
   const balancesAll = useCall<DeriveBalancesAll>(api.derive.balances?.all, [account]);
   const kusamaBalancesAll = useCall<DeriveBalancesAll>(kusamaApi?.derive.balances?.all, [account]);
+  const [freeBalance, setFreeBalance] = useState<BN>();
+  const [freeKusamaBalance, setFreeKusamaBalance] = useState<BN>();
 
   useEffect(() => {
-    void getUserDeposit();
+    // available balance used as free (transferable)
+    if (balancesAll) {
+      setFreeBalance(balancesAll.availableBalance);
+    }
+  }, [balancesAll]);
+
+  useEffect(() => {
+    // available balance used as free (transferable)
+    if (kusamaBalancesAll) {
+      setFreeKusamaBalance(kusamaBalancesAll.availableBalance);
+    }
+  }, [kusamaBalancesAll]);
+
+  useEffect(() => {
+    getUserDeposit && getUserDeposit();
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [balancesAll, kusamaBalancesAll, deposited]);
+  }, [freeBalance, freeKusamaBalance]);
 
   return {
-    balancesAll,
-    deposited,
-    kusamaBalancesAll
+    freeBalance,
+    freeKusamaBalance
   };
 };
