@@ -13,7 +13,7 @@ import envConfig from '@polkadot/apps-config/envConfig';
 import { useApi, useCollection, useFetch } from '@polkadot/react-hooks';
 import { base64Decode, encodeAddress } from '@polkadot/util-crypto';
 
-const { canAddCollections, uniqueCollectionIds } = envConfig;
+const { canAddCollections, uniqueApi, uniqueCollectionIds } = envConfig;
 
 export type MetadataType = {
   metadata?: string;
@@ -106,7 +106,7 @@ export function useCollections () {
    */
   const getOffers = useCallback((page: number, pageSize: number, filters?: Filters) => {
     try {
-      let url = `/offers?page=${page}&pageSize=${pageSize}`;
+      let url = `${uniqueApi}/offers?page=${page}&pageSize=${pageSize}`;
 
       // reset offers before loading first page
       if (page === 1) {
@@ -115,20 +115,23 @@ export function useCollections () {
 
       if (filters) {
         Object.keys(filters).forEach((filterKey: string) => {
-          const currentFilter: string | string[] = filters[filterKey];
+          const currentFilter: string | string[] | number = filters[filterKey];
 
           if (Array.isArray(currentFilter)) {
-            // url += filters[key].map((item: string) => `&collectionId=${item}`).join('');
-            url = `${url}${currentFilter.map((item: string) => `&collectionId=${item}`).join('')}`;
+            if (filterKey === 'collectionIds') {
+              if (!currentFilter?.length) {
+                url = `${url}${envConfig.uniqueCollectionIds.map((item: string) => `&collectionId=${item}`).join('')}`;
+              } else {
+                url = `${url}${currentFilter.map((item: string) => `&collectionId=${item}`).join('')}`;
+              }
+            } else if (filterKey === 'traitsCount') {
+              url = `${url}${currentFilter.map((item: string) => `&traitsCount=${item}`).join('')}`;
+            }
           } else {
-            url += '&' + filterKey + '=' + currentFilter;
+            url += `&${filterKey}=${currentFilter}`;
           }
         });
       }
-
-      // if (!canAddCollections && collectionIds && collectionIds.length) {
-      //   url = `${url}${collectionIds.map((item: string) => `&collectionId=${item}`).join('')}`;
-      // }
 
       fetchData<OffersResponseType>(url).subscribe((result: OffersResponseType | ErrorType) => {
         if (cleanup.current) {
@@ -174,7 +177,7 @@ export function useCollections () {
    */
   const getHoldByMe = useCallback((account: string, page: number, pageSize: number, collectionIds?: string[]) => {
     try {
-      let url = `/OnHold/${account}?page=${page}&pageSize=${pageSize}`;
+      let url = `${uniqueApi}/OnHold/${account}?page=${page}&pageSize=${pageSize}`;
 
       if (!canAddCollections && collectionIds && collectionIds.length) {
         url = `${url}${collectionIds.map((item: string) => `&collectionId=${item}`).join('')}`;
@@ -227,7 +230,7 @@ export function useCollections () {
     page,
     pageSize }: { account?: string, collectionIds?: string[], page: number, pageSize: number }) => {
     try {
-      let url = '/trades';
+      let url = `${uniqueApi}/trades`;
 
       if (account && account.length) {
         url = `${url}/${account}`;
