@@ -8,7 +8,7 @@ import { ApiPromise } from '@polkadot/api/promise';
 import { typesBundle, typesChain } from '@polkadot/apps-config';
 import envConfig from '@polkadot/apps-config/envConfig';
 import { StatusContext } from '@polkadot/react-components';
-import { BalanceInterface } from '@polkadot/react-hooks/useBalance';
+import { useKusamaBalance } from '@polkadot/react-hooks/useKusamaBalance';
 import ApiSigner from '@polkadot/react-signer/signers/ApiSigner';
 import { WsProvider } from '@polkadot/rpc-provider';
 import { TypeRegistry } from '@polkadot/types/create';
@@ -18,10 +18,9 @@ const { kusamaDecimals, minPrice } = envConfig;
 
 interface UseKusamaApiInterface {
   formatKsmBalance: (balance: BN | undefined) => string;
-  getKusamaBalance: () => void;
   getKusamaTransferFee: (recipient: string, value: BN) => Promise<BN | null>;
   kusamaApi: ApiPromise | undefined;
-  kusamaBalance: BalanceInterface | undefined;
+  kusamaAvailableBalance: BN | undefined;
   kusamaTransfer: (recipient: string, value: BN, onSuccess: (status: string) => void, onFail: (status: string) => void) => void;
 }
 
@@ -47,23 +46,11 @@ export function formatStrBalance (decimals: number, value: BN | undefined = new 
 export const useKusamaApi = (account?: string): UseKusamaApiInterface => {
   const { queuePayload, queueSetTxStatus } = useContext(StatusContext);
   const [kusamaApi, setKusamaApi] = useState<ApiPromise>();
-  const [kusamaBalance, setKusamaBalance] = useState<BalanceInterface>();
   const [encodedKusamaAccount, setEncodedKusamaAccount] = useState<string>();
   const { queueExtrinsic } = useContext(StatusContext);
   const kusamaParity = 'wss://kusama-rpc.polkadot.io';
   const kusamaFinality = 'wss://polkadot.api.onfinality.io/public-ws';
-
-  const getKusamaBalance = useCallback(async () => {
-    try {
-      if (kusamaApi && encodedKusamaAccount) {
-        const kusamaAccountBalance: { data: BalanceInterface } = await kusamaApi.query.system.account(encodedKusamaAccount);
-
-        setKusamaBalance(kusamaAccountBalance.data);
-      }
-    } catch (e) {
-      console.log('kusama balance error', e);
-    }
-  }, [encodedKusamaAccount, kusamaApi]);
+  const { kusamaAvailableBalance } = useKusamaBalance(kusamaApi, account);
 
   const initKusamaApi = useCallback((kusamaUrl) => {
     const registry = new TypeRegistry();
@@ -76,6 +63,10 @@ export const useKusamaApi = (account?: string): UseKusamaApiInterface => {
     api.on('ready', (): void => {
       // console.log('kusama api ready');
       setKusamaApi(api);
+    });
+
+    api.on('disconnected', (): void => {
+      window.location.reload();
     });
 
     api.on('error', (): void => {
@@ -117,20 +108,15 @@ export const useKusamaApi = (account?: string): UseKusamaApiInterface => {
   }, [account]);
 
   useEffect(() => {
-    void getKusamaBalance();
-  }, [getKusamaBalance]);
-
-  useEffect(() => {
     initKusamaApi(kusamaParity);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return {
     formatKsmBalance,
-    getKusamaBalance,
     getKusamaTransferFee,
     kusamaApi,
-    kusamaBalance,
+    kusamaAvailableBalance,
     kusamaTransfer
   };
 };
