@@ -1,13 +1,12 @@
 // Copyright 2017-2021 @polkadot/react-params authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type BN from 'bn.js';
 import type { Props } from '../types';
 
-import BN from 'bn.js';
 import React, { useCallback, useMemo } from 'react';
 
 import { Input, InputNumber } from '@polkadot/react-components';
-import { ClassOf } from '@polkadot/types/create';
 import { bnToBn, formatNumber, isUndefined } from '@polkadot/util';
 
 import Bare from './Bare';
@@ -15,7 +14,7 @@ import Bare from './Bare';
 function Amount ({ className = '', defaultValue: { value }, isDisabled, isError, label, onChange, onEnter, registry, type, withLabel }: Props): React.ReactElement<Props> {
   const defaultValue = useMemo(
     () => isDisabled
-      ? value instanceof ClassOf(registry, 'AccountIndex')
+      ? value instanceof registry.createClass('AccountIndex')
         ? value.toString()
         : formatNumber(value as number)
       : bnToBn((value as number) || 0).toString(),
