@@ -6,7 +6,6 @@ import type { Props } from '../types';
 import React, { useCallback } from 'react';
 
 import { Dropdown } from '@polkadot/react-components';
-import { ClassOf } from '@polkadot/types/create';
 import { bnToBn } from '@polkadot/util';
 
 import Bare from './Bare';
@@ -35,7 +34,7 @@ function VoteThresholdParam ({ className = '', defaultValue: { value }, isDisabl
     [onChange]
   );
 
-  const defaultValue = value instanceof ClassOf(registry, 'VoteThreshold')
+  const defaultValue = value instanceof registry.createClass('VoteThreshold')
     ? value.toNumber()
     : bnToBn(value as number).toNumber();
 
